Makefile: rework to avoid wildcards on _repos
The repo collection under _repos might not exist yet when the Makefile
is invoked, so wildcards matching on files under _repos will be empty,
leading to failure. Rework the Makefile to avoid such wildcards by
either matching on files in the docsite repo, which are guaranteed to
exist, or by providing a static file list.

Static file lists where necessary only for libsel4vm
and libsel4vmmplatsupport, which are manually linked to already in the
docsite, so having to update the Makefile when new files are included is
not a big additional burden and will not be easily overlooked.

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>