Add copy button to code block with language set
This only works for code blocks with a declared language, because those
are wrapped in a <div> and we need that diff to anchor the button.

Putting the button into <pre> would work theoretically, but triggers a
bug in Safari when he <pre> block is too wide (scrolling) and innerHTML
of the button is being set (button text disappears).

Signed-off-by: Gerwin Klein <gerwin.klein@proofcraft.systems>